axdevice

Struct AxVmDevices

pub struct AxVmDevices { /* private fields */ }

represent A vm own devices

Implementations

impl AxVmDevices

The implemention for AxVmDevices

pub fn new(config: AxVmDeviceConfig) -> Self

According AxVmDeviceConfig to init the AxVmDevices

pub fn alloc_ivc_channel(&self, size: usize) -> AxResult<GuestPhysAddr>

Allocates an IVC (Inter-VM Communication) channel of the specified size.

pub fn release_ivc_channel(&self, addr: GuestPhysAddr, size: usize) -> AxResult

Releases an IVC channel at the specified address and size.

pub fn add_mmio_dev(&mut self, dev: Arc<dyn BaseMmioDeviceOps>)

Add a MMIO device to the device list

pub fn add_sys_reg_dev(&mut self, dev: Arc<dyn BaseSysRegDeviceOps>)

Add a system register device to the device list

pub fn add_port_dev(&mut self, dev: Arc<dyn BasePortDeviceOps>)

Add a port device to the device list

pub fn iter_mmio_dev(&self) -> impl Iterator<Item = &Arc<dyn BaseMmioDeviceOps>>

Iterates over the MMIO devices in the set.

pub fn iter_sys_reg_dev( &self, ) -> impl Iterator<Item = &Arc<dyn BaseSysRegDeviceOps>>

Iterates over the system register devices in the set.

pub fn iter_port_dev(&self) -> impl Iterator<Item = &Arc<dyn BasePortDeviceOps>>

Iterates over the port devices in the set.

pub fn iter_mut_mmio_dev( &mut self, ) -> impl Iterator<Item = &mut Arc<dyn BaseMmioDeviceOps>>

Iterates over the MMIO devices in the set.

pub fn iter_mut_sys_reg_dev( &mut self, ) -> impl Iterator<Item = &mut Arc<dyn BaseSysRegDeviceOps>>

Iterates over the system register devices in the set.

pub fn iter_mut_port_dev( &mut self, ) -> impl Iterator<Item = &mut Arc<dyn BasePortDeviceOps>>

Iterates over the port devices in the set.

pub fn find_mmio_dev( &self, ipa: GuestPhysAddr, ) -> Option<Arc<dyn BaseMmioDeviceOps>>

Find specific MMIO device by ipa

pub fn find_sys_reg_dev( &self, sys_reg_addr: SysRegAddr, ) -> Option<Arc<dyn BaseSysRegDeviceOps>>

Find specific system register device by ipa

pub fn find_port_dev(&self, port: Port) -> Option<Arc<dyn BasePortDeviceOps>>

Find specific port device by port number

pub fn handle_mmio_read( &self, addr: GuestPhysAddr, width: AccessWidth, ) -> AxResult<usize>

Handle the MMIO read by GuestPhysAddr and data width, return the value of the guest want to read

pub fn handle_mmio_write( &self, addr: GuestPhysAddr, width: AccessWidth, val: usize, ) -> AxResult

Handle the MMIO write by GuestPhysAddr, data width and the value need to write, call specific device to write the value

pub fn handle_sys_reg_read( &self, addr: SysRegAddr, width: AccessWidth, ) -> AxResult<usize>

Handle the system register read by SysRegAddr and data width, return the value of the guest want to read

pub fn handle_sys_reg_write( &self, addr: SysRegAddr, width: AccessWidth, val: usize, ) -> AxResult

Handle the system register write by SysRegAddr, data width and the value need to write, call specific device to write the value

pub fn handle_port_read( &self, port: Port, width: AccessWidth, ) -> AxResult<usize>

Handle the port read by port number and data width, return the value of the guest want to read

pub fn handle_port_write( &self, port: Port, width: AccessWidth, val: usize, ) -> AxResult

Handle the port write by port number, data width and the value need to write, call specific device to write the value